feat: Implement `WelcomeChatContainer` with `ChatContainer`
