+++ src/containers/ChatContainer/ChatContainer.tsx
@@ -1,6 +1,3 @@
-import {useEffect, useMemo, useRef, type FC} from "react"
-import {assert} from "../../utils/util"
-import IconButton from "../../components/IconButton"
 import TypingIndicator from "../../components/TypingIndicator"
 import useActiveRoom, {MessageKind} from "@/hooks/matrix/useActiveRoom"
 import ImageMessage from "../../components/ImageMessage"
@@ -9,17 +6,13 @@ import EventMessage from "../../components/EventMessage"
 import {twMerge} from "tailwind-merge"
 import Button, {ButtonVariant} from "../../components/Button"
 import UnreadIndicator from "../../components/UnreadIndicator"
-import {
-  IoEllipsisVertical,
-  IoInformationCircle,
-  IoAttach,
-  IoPaperPlane,
-} from "react-icons/io5"
-import {IoIosHappy, IoMdLink} from "react-icons/io"
-import {LiaSlackHash} from "react-icons/lia"
 import Modal, {ModalRenderLocation} from "../../components/Modal"
 import SmartActionBar from "@/containers/SmartActionBar"
 import useChatInput from "./useChatInput"
+import ChatHeader from "./ChatHeader"
+import ChatInput from "./ChatInput"
+import {useMemo, type FC} from "react"
+import WelcomeChatContainer from "./WelcomeChatContainer"
 
 export type ChatContainerProps = {
   className?: string
@@ -37,6 +30,7 @@ const ChatContainer: FC<ChatContainerProps> = ({className}) => {
     filesContent,
     clear,
     openFilePicker,
+    activeRoomId,
   } = useActiveRoom()
 
   const messages = useMemo(
@@ -82,170 +76,60 @@ const ChatContainer: FC<ChatContainerProps> = ({className}) => {
         </div>
       </Modal>
 
-      <div
-        id={ModalRenderLocation.ChatContainer}
-        className={twMerge(
-          "flex h-screen flex-col gap-4 border border-stone-200",
-          className
-        )}>
-        <ChatHeader roomName={roomName} />
-
+      {activeRoomId === null ? (
+        <WelcomeChatContainer />
+      ) : (
         <div
-          ref={scrollReference => {
-            if (scrollReference === null) {
-              return
-            }
+          id={ModalRenderLocation.ChatContainer}
+          className={twMerge(
+            "flex h-screen flex-col gap-4 border border-stone-200",
+            className
+          )}>
+          <ChatHeader roomName={roomName} />
 
-            scrollReference.scrollTo({
-              top: scrollReference.scrollHeight - scrollReference.clientHeight,
-              behavior: "smooth",
-            })
-          }}
-          className="mx-4 flex max-h-full grow flex-col gap-4 overflow-y-auto scroll-smooth scrollbar-hide">
-          {messages}
-        </div>
+          <div
+            ref={scrollReference => {
+              if (scrollReference === null) {
+                return
+              }
 
-        <div className="mx-4 flex flex-col gap-3">
-          <ChatInput
-            onAttach={openFilePicker}
-            onSend={() => {
-              void sendTextMessage(messageText)
-
-              setMessageText("")
+              scrollReference.scrollTo({
+                top:
+                  scrollReference.scrollHeight - scrollReference.clientHeight,
+                behavior: "smooth",
+              })
             }}
-            onValueChange={setMessageText}
-            value={messageText}
-          />
-
-          <div className="flex gap-3">
-            <div className="size-6" />
-
-            <div className="size-6" />
-
-            {typingUsers.length > 0 && <TypingIndicator users={typingUsers} />}
+            className="mx-4 flex max-h-full grow flex-col gap-4 overflow-y-auto scroll-smooth scrollbar-hide">
+            {messages}
           </div>
-        </div>
 
-        <SmartActionBar />
-      </div>
+          <div className="mx-4 flex flex-col gap-3">
+            <ChatInput
+              onAttach={openFilePicker}
+              onSend={() => {
+                void sendTextMessage(messageText)
+
+                setMessageText("")
+              }}
+              onValueChange={setMessageText}
+              value={messageText}
+            />
+
+            <div className="flex gap-3">
+              <div className="size-6" />
+
+              <div className="size-6" />
+
+              {typingUsers.length > 0 && (
+                <TypingIndicator users={typingUsers} />
+              )}
+            </div>
+          </div>
+
+          <SmartActionBar />
+        </div>
+      )}
     </>
-  )
-}
-
-type ChatInputProps = {
-  value: string
-  onAttach: () => void
-  onSend: () => void
-  onValueChange: (value: string) => void
-}
-
-const ChatInput: FC<ChatInputProps> = ({
-  onAttach,
-  onSend,
-  onValueChange,
-  value,
-}) => {
-  const textareaReference = useRef<HTMLTextAreaElement>(null)
-
-  const handleKeyDown = (event: React.KeyboardEvent) => {
-    if (event.key === "Enter")
-      if (event.ctrlKey) {
-        onValueChange(value + "\n")
-      } else {
-        event.preventDefault()
-        onSend()
-      }
-  }
-
-  useEffect(() => {
-    // Move scroll to last message.
-    const textarea = textareaReference.current
-
-    if (textarea !== null) {
-      textarea.style.height = "auto"
-      textarea.style.height = `${textarea.scrollHeight}px`
-    }
-  }, [value])
-
-  return (
-    <div className="flex gap-2">
-      <div className="flex h-full items-center gap-3">
-        <IconButton
-          onClick={() => {
-            /* TODO: Handle `emoji` button click. */
-          }}
-          tooltip="Emoji"
-          Icon={IoIosHappy}
-        />
-
-        <IconButton onClick={onAttach} tooltip="Attach" Icon={IoAttach} />
-      </div>
-
-      <div className="flex w-full rounded-md border border-neutral-300 bg-neutral-50">
-        <textarea
-          onKeyDown={handleKeyDown}
-          rows={1}
-          ref={textareaReference}
-          placeholder="Write a message or simply say 👋🏼 hello..."
-          value={value}
-          disabled={false}
-          onChange={value => {
-            onValueChange(value.target.value)
-          }}
-          className="flex max-h-24 w-full resize-none overflow-y-auto border-none bg-transparent p-3 scrollbar-hide focus-visible:outline-none focus-visible:outline-0"
-        />
-
-        <div className="m-1 size-max">
-          <IconButton
-            tooltip="Send"
-            Icon={IoPaperPlane}
-            color="#C463FF"
-            isDisabled={false}
-            onClick={onSend}
-          />
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const ChatHeader: FC<{roomName: string}> = ({roomName}) => {
-  assert(roomName.length > 0, "room name should not be empty")
-
-  return (
-    <header className="flex items-center gap-2 border-b border-b-stone-200 p-3">
-      <div className="m-2 flex w-full gap-1">
-        <LiaSlackHash className="text-purple-500" />
-
-        <span className="text-purple-500">{roomName}</span>
-
-        {/* <span className="text-stone-600">{text}</span> */}
-      </div>
-
-      <IconButton
-        onClick={() => {
-          /* TODO: Handle `info` button click. */
-        }}
-        tooltip="Room details"
-        Icon={IoInformationCircle}
-      />
-
-      <IconButton
-        onClick={() => {
-          /* TODO: Handle `link` button click. */
-        }}
-        tooltip="Copy link"
-        Icon={IoMdLink}
-      />
-
-      <IconButton
-        onClick={() => {
-          /* TODO: Handle `more` button click. */
-        }}
-        tooltip="More actions"
-        Icon={IoEllipsisVertical}
-      />
-    </header>
   )
 }
